requirements.txt
Django>=4.2,<6
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.1
requests>=2.31
django-cors-headers>=4.3
